@@ -1,41 +1,36 @@
 {
-  "name": "aos",
-  "version": "3.0.0-beta.6",
+  "name": "@ignitekit/aos",
+  "version": "3.0.0",
+  "type": "module",
   "description": "Animate on scroll library",
-  "homepage": "https://michalsnik.github.io/aos/",
-  "author": "Michał Sajnóg <michal.sajnog@hotmail.com>",
+  "homepage": "https://github.com/ignitekit/aos",
+  "author": "IgniteKit <ignitekit@ideologix.com>",
   "license": "MIT",
   "main": "dist/aos.cjs.js",
   "module": "dist/aos.esm.js",
   "browser": "dist/aos.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/michalsnik/aos.git"
+    "url": "https://github.com/ignitekit/aos.git"
   },
   "bugs": {
-    "url": "https://github.com/michalsnik/aos/issues"
+    "url": "https://github.com/ignitekit/aos/issues"
   },
   "devDependencies": {
-    "autoprefixer": "^8.4.1",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.3",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "cssnano": "^3.10.0",
-    "cypress": "^2.1.0",
-    "eslint": "^4.19.1",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.0",
-    "live-server": "tapio/live-server#master",
-    "node-sass": "^4.9.0",
-    "npm-run-all": "^4.1.3",
-    "rollup": "^0.58.2",
-    "rollup-plugin-babel": "^3.0.4",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-postcss": "^1.6.1",
-    "rollup-plugin-uglify": "^3.0.0"
+    "autoprefixer": "^10.4.20",
+    "cssnano": "^7.0.6",
+    "cypress": "^13.17.0",
+    "eslint": "^9.18.0",
+    "express": "^4.21.2",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.14.0",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.49",
+    "prettier": "^3.4.2",
+    "sass": "^1.83.0",
+    "terser": "^5.37.0",
+    "vite": "^6.0.7"
   },
   "dependencies": {
     "classlist-polyfill": "^1.2.0",
@@ -43,17 +38,19 @@
     "lodash.throttle": "^4.1.1"
   },
   "scripts": {
-    "build": "NODE_ENV=production rollup -c",
-    "watch": "NODE_ENV=dev rollup -c -w",
-    "serve": "node ./scripts/start-server.js",
-    "dev": "npm-run-all --parallel serve watch",
-    "test": "yarn lint && NODE_ENV=test node ./scripts/run-cypress-tests.js",
+    "build": "vite build",
+    "dev": "vite",
+    "preview": "vite preview",
+    "test": "npm run lint && node ./scripts/run-cypress-tests.js",
     "test:dev": "cypress open",
-    "lint": "eslint src cypress demo scripts",
+    "lint": "eslint src cypress scripts",
     "prepare": "npm run build"
   },
   "files": [
     "dist",
     "src"
-  ]
+  ],
+  "publishConfig": {
+    "access": "public"
+  }
 }
