@@ -1,15 +1,15 @@
-describe('mutation observer', function() {
+describe('mutation observer', function () {
   before(() => {
     cy.visit('/async.html');
     cy.initAOS();
   });
 
-  it('Should not animate any items if not present', function() {
+  it('Should not animate any items if not present', function () {
     cy.get('.aos-item').should('have.length', 0);
     cy.get('.aos-animate').should('have.length', 0);
   });
 
-  it('Should animate new items as they appear', function() {
+  it('Should animate new items as they appear', function () {
     cy.dispatchEvent('add-aos-item');
     cy.get('.aos-item').should('have.length', 1);
     cy.get('.aos-animate').should('have.length', 1);
@@ -46,7 +46,7 @@ describe('mutation observer', function() {
     cy.get('.aos-animate').should('have.length', 6);
   });
 
-  it('Should gracefully handle nodes deletion', function() {
+  it('Should gracefully handle nodes deletion', function () {
     Cypress.$('.aos-item')[1].remove();
     cy.get('.aos-item').should('have.length', 23);
     cy.get('.aos-animate').should('have.length', 6);
