@@ -1,13 +1,13 @@
-describe('setting: animatedClassName', function() {
-  context('with: "rawr"', function() {
+describe('setting: animatedClassName', function () {
+  context('with: "rawr"', function () {
     before(() => {
       cy.visit('/');
       cy.initAOS({
-        animatedClassName: 'rawr'
+        animatedClassName: 'rawr',
       });
     });
 
-    it('Should set proper class names on AOS elements', function() {
+    it('Should set proper class names on AOS elements', function () {
       cy.get('.aos-animate').should('have.length', 0);
       cy.get('.rawr').should('have.length', 6);
 
@@ -17,15 +17,15 @@ describe('setting: animatedClassName', function() {
     });
   });
 
-  context('with: null', function() {
+  context('with: null', function () {
     before(() => {
       cy.visit('/');
       cy.initAOS({
-        animatedClassName: null
+        animatedClassName: null,
       });
     });
 
-    it('Should not set class names on AOS elements on scroll', function() {
+    it('Should not set class names on AOS elements on scroll', function () {
       cy.get('.aos-animate').should('have.length', 0);
       cy.scrollTo(0, 800);
       cy.get('.aos-animate').should('have.length', 0);
