@@ -1,11 +1,11 @@
-describe('setting: disable', function() {
+describe('setting: disable', function () {
   beforeEach(() => {
     cy.visit('/');
   });
 
-  it('Should properly disable AOS', function() {
+  it('Should properly disable AOS', function () {
     cy.initAOS({
-      disable: true
+      disable: true,
     });
 
     cy.get('.aos-item[data-aos]').should('have.length', 0);
@@ -13,12 +13,12 @@ describe('setting: disable', function() {
     cy.get('.aos-animate').should('have.length', 0);
   });
 
-  it('Should respect function passed as "disable" value', function() {
+  it('Should respect function passed as "disable" value', function () {
     cy.viewport(360, 420);
 
     cy.window().then(({ AOS, innerWidth }) => {
       AOS.init({
-        disable: () => innerWidth < 400
+        disable: () => innerWidth < 400,
       });
     });
 
