@@ -1,13 +1,13 @@
-describe('setting: useClassNames', function() {
-  context('just useClassNames', function() {
+describe('setting: useClassNames', function () {
+  context('just useClassNames', function () {
     before(() => {
       cy.visit('/animatecss.html');
       cy.initAOS({
-        useClassNames: true
+        useClassNames: true,
       });
     });
 
-    it('Should set proper custom class names on AOS elements', function() {
+    it('Should set proper custom class names on AOS elements', function () {
       cy.get('.aos-animate').should('have.length', 6);
       cy.get('.fadeInUp').should('have.length', 6);
 
@@ -17,17 +17,17 @@ describe('setting: useClassNames', function() {
     });
   });
 
-  context('with animatedClassName and initClassName', function() {
+  context('with animatedClassName and initClassName', function () {
     before(() => {
       cy.visit('/animatecss.html');
       cy.initAOS({
         useClassNames: true,
         initClassName: false,
-        animatedClassName: 'animate'
+        animatedClassName: 'animate',
       });
     });
 
-    it('Should set proper custom class names on AOS elements', function() {
+    it('Should set proper custom class names on AOS elements', function () {
       cy.get('.aos-init').should('have.length', 0);
       cy.get('.aos-animate').should('have.length', 0);
 
