@@ -3,7 +3,7 @@
 import { getPositionIn, getPositionOut } from './offsetCalculator';
 import getInlineOption from './getInlineOption';
 
-const prepare = function($elements, options) {
+const prepare = function ($elements, options) {
   $elements.forEach((el, i) => {
     const mirror = getInlineOption(el.node, 'mirror', options.mirror);
     const once = getInlineOption(el.node, 'once', options.once);
@@ -21,14 +21,14 @@ const prepare = function($elements, options) {
 
     el.position = {
       in: getPositionIn(el.node, options.offset, options.anchorPlacement),
-      out: mirror && getPositionOut(el.node, options.offset)
+      out: mirror && getPositionOut(el.node, options.offset),
     };
 
     el.options = {
       once,
       mirror,
       animatedClassNames,
-      id
+      id,
     };
   });
 
